# docs/swagger.yaml
basePath: /
definitions:
...
  apicommon.IntegratorInfoResponse:
    properties:
      enabled:
        type: boolean
      limits:
        $ref: '#/definitions/apicommon.IntegratorLimits'
      usage:
        $ref: '#/definitions/apicommon.IntegratorUsage'
    type: object
...
  apicommon.ListManagedOrganizations:
    properties:
      organizations:
        items:
          $ref: '#/definitions/apicommon.OrganizationInfo'
        type: array
      pagination:
        $ref: '#/definitions/apicommon.Pagination'
    type: object
...
  apicommon.OrganizationInfo:
    properties:
      active:
        description: Whether the organization is active
        type: boolean
      address:
        description: The organization's blockchain address
        items:
          type: integer
        type: array
      color:
        description: The organization's brand color in hex format
        type: string
      communications:
        description: Whether the organization has enabled communications
        type: boolean
      counters:
        allOf:
        - $ref: '#/definitions/apicommon.SubscriptionUsage'
        description: Usage counters for the organization
      country:
        description: The country where the organization is based
        type: string
      createdAt:
        description: Creation timestamp in RFC3339 format
        type: string
      description:
        allOf:
        - $ref: '#/definitions/apicommon.MultilingualText'
        description: Description is a shorthand for meta["description"]. Same encoding
          rules as Name.
      integrator:
        description: |-
          Whether to subscribe the new organization to the free integrator plan at
          creation time (opt-in). Used by the integrator portal so a newly created org
          becomes an integrator on the free tier with no checkout. Default false uses
          the regular default plan.
        type: boolean
      logo:
        allOf:
        - $ref: '#/definitions/apicommon.MultilingualText'
        description: Logo is a shorthand for meta["logo"]. Same encoding rules as
          Name.
      managedBy:
        description: |-
          Integrator organization that manages this org, when it was created through
          the integrator portal. Nil for regular organizations. Lets clients tell
          managed orgs apart from the user's own orgs (e.g. to hide them from the org
          switcher), a distinction Parent does not capture (managed orgs set ManagedBy,
          not Parent).

          A pointer so it is omitted rather than emitted as the zero address: json's
          omitempty does not skip the fixed-size common.Address array.
        example: deadbeef
        format: hex
        type: string
      meta:
        additionalProperties: {}
        description: Arbitrary key value fields with metadata regarding the organization
        type: object
      name:
        allOf:
        - $ref: '#/definitions/apicommon.MultilingualText'
        description: |-
          Name is a shorthand for meta["name"]. On write accepts a plain string
          (stored as {"default": "<string>"}) or a locale map; on read it mirrors
          whatever is stored in meta["name"]. If both Name and meta["name"] are
          provided on a create request, meta["name"] takes precedence.
      parent:
        allOf:
        - $ref: '#/definitions/apicommon.OrganizationInfo'
        description: Parent organization if this is a sub-organization
      size:
        description: The size category of the organization
        type: string
      subdomain:
        description: The organization's subdomain
        type: string
      subscription:
        allOf:
        - $ref: '#/definitions/apicommon.SubscriptionDetails'
        description: Subscription details for the organization
      timezone:
        description: The organization's timezone
        type: string
      type:
        description: The type of organization
        type: string
      website:
        description: The organization's website URL
        type: string
    type: object
...
  errors.Error:
    properties:
      code:
        description: Error code
        type: integer
      data:
        description: Optional data to include in the error response
      error:
        description: Original error
      httpstatus:
        description: HTTP status code to return
        type: integer
      logLevel:
        description: Log level for this error (defaults to "debug")
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: info@vocdoni.io
    name: API Support
    url: https://vocdoni.io
  description: API for Vocdoni SaaS Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vocdoni SaaS API
  version: "1.0"
paths:
...
  /integrator:
    get:
      description: |-
        Returns whether the caller's organization (resolved from the API key's org or the user
        session — no address in the URL) is enabled as an integrator, along with its effective
        managed-resource limits and current usage. The caller must be an admin or manager of the
        organization. When the organization is not an integrator, `enabled` is false and `limits`
        is omitted (usage counters are still returned). A 0 in any limit field means unlimited.

        Also callable with a scoped API key (scope: `quota:read`).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicommon.IntegratorInfoResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Get integrator quota and usage
      tags:
      - integrator
  /integrator/organizations:
    get:
      description: |-
        Returns a paginated list of organizations managed by the caller's integrator
        organization (resolved from the API key's org or the user session — no address in the
        URL). The caller must be an admin or manager of the integrator organization.

        Also callable with a scoped API key (scope: `managed:read`).
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicommon.ListManagedOrganizations'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: List organizations managed by an integrator
      tags:
      - integrator
    post:
      consumes:
      - application/json
      description: |-
        Create a new organization managed by the caller's integrator organization. The
        integrator is resolved from the authenticated principal (the API key's org, or the
        user's organization for a session) — no address is passed in the URL. The caller must
        be an admin of that integrator organization, which must be enabled as an integrator and
        within its managed-organizations quota. The managed org's on-chain account is always
        provisioned eagerly. The optional `ownerEmail` designates the managed org's owner/admin
        (defaults to the calling user); the per-user MaxOrgsPerUser cap is bypassed for managed orgs.

        Also callable with a scoped API key (scope: `managed:write`).
      parameters:
      - description: Managed organization information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apicommon.CreateManagedOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicommon.OrganizationInfo'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Not an integrator or quota reached
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Integrator organization not found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Create a managed organization under an integrator
      tags:
      - integrator
  /integrator/organizations/{orgAddress}:
    delete:
      description: |-
        Delete the managed organization at {orgAddress} together with every piece of data
        tied to it (memberbase, censuses + participants, processes, bundles, CSP tokens,
        jobs, pending invites), and unlink it from its members. The caller must be an admin
        of their integrator organization (resolved from the API key's org or the user session
        — no address in the URL), and the target must be managed by it. The integrator's usage
        counters are rolled back accordingly.

        Deletion is blocked with 409 when any of the managed org's published elections is
        still active on-chain (READY/PAUSED); end them (or wait for them to end) first.
        The on-chain organization account and any published elections/censuses are immutable
        on the Vochain and are not affected — only DB-side data is removed.

        Also callable with a scoped API key (scope: `managed:write`).
      parameters:
      - description: Managed organization address to delete
        in: path
        name: orgAddress
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicommon.DeleteManagedOrganizationResponse'
        "400":
          description: Invalid address
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Managed organization not found
          schema:
            $ref: '#/definitions/errors.Error'
        "409":
          description: Managed organization has active elections
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - BearerAuth: []
      summary: Delete a managed organization and all its data
      tags:
      - integrator
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: |-
      Type "Bearer" followed by a space and either a JWT (user session) or a
      scoped API key (prefixed "vsk_"). API keys are accepted only on endpoints
      that opt into key auth, and only when the key carries the required scope.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"